requirements.txt
Flask
numpy
opencv-python
albumentations
torch
Pillow
scikit-image
iglovikov-helper-functions
segmentation-models-pytorch
torchvision
pytesseract
midv500models